pyproject.toml
[tool.poetry]
name = "store-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Kiel Ed"]
readme = "README.md"

[tool.black]
line-length = 88
target-version = [ "py310" ]
exclude = "((.eggs | .git | .pytest_cache | build | dist | */migrations))"

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
]
ignore = [
    "B904",
    "B006",
    "E501",
    "B008",
    "C901",
    "F403",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.poetry.dependencies]
python = "^3.11.3"
alembic = "^1.10.3"
fastapi = "^0.95.0"
python-telegram-bot = { extras = ["webhooks"], version = "^20.2" }
pydantic = "^1.10.7"
sqlalchemy = "^2.0.9"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
aiohttp = { extras = ["speedups"], version = "^3.8.4" }
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
sqladmin = "^0.10.1"

[tool.poetry.dev-dependencies]
pre-commit = "^3.2.2"
pytest = "^7.3.0"
black = "^23.3.0"
ruff = "^0.0.261"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
